CID 351720 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument db_cache.size() to format specifier %d was expected to have type int but has type unsigned long

+++ src/world/Management/LootMgr.cpp
@@ -334,7 +334,7 @@ void LootMgr::LoadLootTables(const char* szTableName, LootStore* LootTable)
             (*LootTable)[entry_id] = list;
         }
     }
-    LogDetail("%d loot templates loaded from %s", db_cache.size(), szTableName);
+    LogDetail("%u loot templates loaded from %s", static_cast<uint32_t>(db_cache.size()), szTableName);
     delete result;
 }
